Move StreamLayerClient online test to separate file
Move online tests from TestStreamLayerClient to
StreamLayerClientOnlineTest.

Relates-to: OLPEDGE-720
