+++ olp-cpp-sdk-dataservice-write/tests/TestStreamLayerClient.cpp
@@ -205,560 +205,6 @@ class StreamLayerClientTestBase : public ::testing::TestWithParam<bool> {
   std::shared_ptr<std::vector<unsigned char>> sdii_data_;
 };
 
-class StreamLayerClientOnlineTest : public StreamLayerClientTestBase {
- protected:
-  static std::shared_ptr<olp::http::Network> s_network;
-
-  static void SetUpTestSuite() {
-    s_network = olp::client::OlpClientSettingsFactory::
-        CreateDefaultNetworkRequestHandler();
-  }
-
-  virtual std::shared_ptr<StreamLayerClient> CreateStreamLayerClient()
-      override {
-    auto network = s_network;
-
-    olp::authentication::Settings authentication_settings;
-    authentication_settings.token_endpoint_url =
-        CustomParameters::getArgument(kEndpoint);
-    authentication_settings.network_request_handler = network;
-
-    olp::authentication::TokenProviderDefault provider(
-        CustomParameters::getArgument(kAppid),
-        CustomParameters::getArgument(kSecret), authentication_settings);
-
-    olp::client::AuthenticationSettings auth_client_settings;
-    auth_client_settings.provider = provider;
-
-    olp::client::OlpClientSettings settings;
-    settings.authentication_settings = auth_client_settings;
-    settings.network_request_handler = network;
-    settings.task_scheduler =
-        olp::client::OlpClientSettingsFactory::CreateDefaultTaskScheduler(1u);
-
-    return std::make_shared<StreamLayerClient>(
-        olp::client::HRN{GetTestCatalog()}, settings);
-  }
-};
-
-// Static network instance is necessary as it needs to outlive any created
-// clients. This is a known limitation as triggered send requests capture the
-// network instance inside the callbacks.
-std::shared_ptr<olp::http::Network> StreamLayerClientOnlineTest::s_network;
-
-INSTANTIATE_TEST_SUITE_P(TestOnline, StreamLayerClientOnlineTest,
-                         ::testing::Values(true));
-
-TEST_P(StreamLayerClientOnlineTest, PublishData) {
-  auto response =
-      client_
-          ->PublishData(
-              PublishDataRequest().WithData(data_).WithLayerId(GetTestLayer()))
-          .GetFuture()
-          .get();
-
-  ASSERT_NO_FATAL_FAILURE(PublishDataSuccessAssertions(response));
-}
-
-TEST_P(StreamLayerClientOnlineTest, PublishDataGreaterThanTwentyMib) {
-  auto large_data =
-      std::make_shared<std::vector<unsigned char>>(kTwentyMib + 1, 'z');
-
-  auto response = client_
-                      ->PublishData(PublishDataRequest()
-                                        .WithData(large_data)
-                                        .WithLayerId(GetTestLayer()))
-                      .GetFuture()
-                      .get();
-
-  ASSERT_NO_FATAL_FAILURE(PublishDataSuccessAssertions(response));
-}
-
-TEST_P(StreamLayerClientOnlineTest, PublishDataAsync) {
-  std::promise<PublishDataResponse> response_promise;
-  bool call_is_async = true;
-
-  auto cancel_token = client_->PublishData(
-      PublishDataRequest().WithData(data_).WithLayerId(GetTestLayer()),
-      [&](const PublishDataResponse& response) {
-        call_is_async = false;
-        response_promise.set_value(response);
-      });
-
-  EXPECT_TRUE(call_is_async);
-  auto response_future = response_promise.get_future();
-  auto status = response_future.wait_for(std::chrono::seconds(30));
-  if (status != std::future_status::ready) {
-    cancel_token.cancel();
-  }
-  auto response = response_future.get();
-
-  ASSERT_NO_FATAL_FAILURE(PublishDataSuccessAssertions(response));
-}
-
-TEST_P(StreamLayerClientOnlineTest, PublishDataCancel) {
-  auto cancel_future = client_->PublishData(
-      PublishDataRequest().WithData(data_).WithLayerId(GetTestLayer()));
-
-  std::thread([cancel_future]() {
-    std::this_thread::sleep_for(std::chrono::milliseconds(200));
-    cancel_future.GetCancellationToken().cancel();
-  }).detach();
-
-  auto response = cancel_future.GetFuture().get();
-
-  // If the response is successful, do not fail to avoid flakiness. This could
-  // be becauase of a fast network for example.
-  if (response.IsSuccessful()) {
-    SUCCEED();
-    return;
-  }
-
-  ASSERT_NO_FATAL_FAILURE(PublishFailureAssertions(response));
-  // TODO OlpClient/ API level should be updated to fill in the proper APIError
-  // from HttpResponse?
-  //  ASSERT_EQ(olp::client::ErrorCode::Cancelled,
-  //  response.GetError().GetErrorCode());
-}
-
-TEST_P(StreamLayerClientOnlineTest, PublishDataCancelLongDelay) {
-  auto cancel_future = client_->PublishData(
-      PublishDataRequest().WithData(data_).WithLayerId(GetTestLayer()));
-
-  std::thread([cancel_future]() {
-    std::this_thread::sleep_for(std::chrono::milliseconds(1200));
-    cancel_future.GetCancellationToken().cancel();
-  }).detach();
-
-  auto response = cancel_future.GetFuture().get();
-
-  // If the response is successful, do not fail to avoid flakiness. This could
-  // be becauase of a fast network for example.
-  if (response.IsSuccessful()) {
-    SUCCEED();
-    return;
-  }
-
-  ASSERT_NO_FATAL_FAILURE(PublishFailureAssertions(response));
-  // TODO OlpClient/ API level should be updated to fill in the proper APIError
-  // from HttpResponse?
-  //  ASSERT_EQ(olp::client::ErrorCode::Cancelled,
-  //  response.GetError().GetErrorCode());
-}
-
-TEST_P(StreamLayerClientOnlineTest,
-       PublishDataCancelGetFutureAfterRequestCancelled) {
-  auto cancel_future = client_->PublishData(
-      PublishDataRequest().WithData(data_).WithLayerId(GetTestLayer()));
-
-  std::thread([cancel_future]() {
-    std::this_thread::sleep_for(std::chrono::milliseconds(200));
-    cancel_future.GetCancellationToken().cancel();
-  }).detach();
-
-  std::this_thread::sleep_for(std::chrono::milliseconds(400));
-  auto response = cancel_future.GetFuture().get();
-
-  // If the response is successful, do not fail to avoid flakiness. This could
-  // be becauase of a fast network for example.
-  if (response.IsSuccessful()) {
-    SUCCEED();
-    return;
-  }
-
-  ASSERT_NO_FATAL_FAILURE(PublishFailureAssertions(response));
-  // TODO OlpClient/ API level should be updated to fill in the proper APIError
-  // from HttpResponse?
-  //  ASSERT_EQ(olp::client::ErrorCode::Cancelled,
-  //  response.GetError().GetErrorCode());
-}
-
-TEST_P(StreamLayerClientOnlineTest, PublishDataGreaterThanTwentyMibCancel) {
-  auto large_data =
-      std::make_shared<std::vector<unsigned char>>(kTwentyMib + 1, 'z');
-
-  auto cancel_future = client_->PublishData(
-      PublishDataRequest().WithData(large_data).WithLayerId(GetTestLayer()));
-
-  std::thread([cancel_future]() {
-    std::this_thread::sleep_for(std::chrono::milliseconds(500));
-    cancel_future.GetCancellationToken().cancel();
-  }).detach();
-
-  auto response = cancel_future.GetFuture().get();
-
-  // If the response is successful, do not fail to avoid flakiness. This could
-  // be becauase of a fast network for example.
-  // TODO(EDE-881) remove the need to do this.
-  if (response.IsSuccessful()) {
-    SUCCEED();
-    return;
-  }
-
-  ASSERT_NO_FATAL_FAILURE(PublishFailureAssertions(response));
-}
-
-TEST_P(StreamLayerClientOnlineTest, IncorrectLayer) {
-  auto response =
-      client_
-          ->PublishData(
-              PublishDataRequest().WithData(data_).WithLayerId("BadLayer"))
-          .GetFuture()
-          .get();
-
-  ASSERT_NO_FATAL_FAILURE(PublishFailureAssertions(response));
-}
-
-TEST_P(StreamLayerClientOnlineTest, NullData) {
-  auto response =
-      client_
-          ->PublishData(PublishDataRequest().WithData(nullptr).WithLayerId(
-              GetTestLayer()))
-          .GetFuture()
-          .get();
-
-  ASSERT_NO_FATAL_FAILURE(PublishFailureAssertions(response));
-}
-
-TEST_P(StreamLayerClientOnlineTest, CustomTraceId) {
-  const auto uuid = GenerateRandomUUID();
-
-  auto response = client_
-                      ->PublishData(PublishDataRequest()
-                                        .WithData(data_)
-                                        .WithLayerId(GetTestLayer())
-                                        .WithTraceId(uuid))
-                      .GetFuture()
-                      .get();
-
-  ASSERT_NO_FATAL_FAILURE(PublishDataSuccessAssertions(response));
-
-  EXPECT_STREQ(response.GetResult().GetTraceID().c_str(), uuid.c_str());
-}
-
-TEST_P(StreamLayerClientOnlineTest, BillingTag) {
-  auto response = client_
-                      ->PublishData(PublishDataRequest()
-                                        .WithData(data_)
-                                        .WithLayerId(GetTestLayer())
-                                        .WithBillingTag(kBillingTag))
-                      .GetFuture()
-                      .get();
-
-  ASSERT_NO_FATAL_FAILURE(PublishDataSuccessAssertions(response));
-}
-
-#ifdef DATASERVICE_WRITE_HAS_OPENSSL
-TEST_P(StreamLayerClientOnlineTest, ChecksumValid) {
-  const std::string data_string(data_->begin(), data_->end());
-  auto checksum = sha256(data_string);
-
-  auto response = client_
-                      ->PublishData(PublishDataRequest()
-                                        .WithData(data_)
-                                        .WithLayerId(GetTestLayer())
-                                        .WithChecksum(checksum))
-                      .GetFuture()
-                      .get();
-
-  ASSERT_NO_FATAL_FAILURE(PublishDataSuccessAssertions(response));
-}
-#endif
-
-TEST_P(StreamLayerClientOnlineTest, ChecksumGarbageString) {
-  auto response = client_
-                      ->PublishData(PublishDataRequest()
-                                        .WithData(data_)
-                                        .WithLayerId(GetTestLayer())
-                                        .WithChecksum("GarbageChecksum"))
-                      .GetFuture()
-                      .get();
-
-  ASSERT_NO_FATAL_FAILURE(PublishFailureAssertions(response));
-}
-
-TEST_P(StreamLayerClientOnlineTest, SequentialPublishSameLayer) {
-  auto response =
-      client_
-          ->PublishData(
-              PublishDataRequest().WithData(data_).WithLayerId(GetTestLayer()))
-          .GetFuture()
-          .get();
-
-  ASSERT_NO_FATAL_FAILURE(PublishDataSuccessAssertions(response));
-
-  response = client_
-                 ->PublishData(PublishDataRequest().WithData(data_).WithLayerId(
-                     GetTestLayer()))
-                 .GetFuture()
-                 .get();
-
-  ASSERT_NO_FATAL_FAILURE(PublishDataSuccessAssertions(response));
-}
-
-TEST_P(StreamLayerClientOnlineTest, SequentialPublishDifferentLayer) {
-  auto response =
-      client_
-          ->PublishData(
-              PublishDataRequest().WithData(data_).WithLayerId(GetTestLayer()))
-          .GetFuture()
-          .get();
-
-  ASSERT_NO_FATAL_FAILURE(PublishDataSuccessAssertions(response));
-
-  response = client_
-                 ->PublishData(PublishDataRequest().WithData(data_).WithLayerId(
-                     GetTestLayer2()))
-                 .GetFuture()
-                 .get();
-
-  ASSERT_NO_FATAL_FAILURE(PublishDataSuccessAssertions(response));
-}
-
-TEST_P(StreamLayerClientOnlineTest, ConcurrentPublishSameIngestApi) {
-  auto publish_data = [&]() {
-    auto response =
-        client_
-            ->PublishData(PublishDataRequest().WithData(data_).WithLayerId(
-                GetTestLayer()))
-            .GetFuture()
-            .get();
-    ASSERT_NO_FATAL_FAILURE(PublishDataSuccessAssertions(response));
-  };
-
-  auto async1 = std::async(std::launch::async, publish_data);
-  auto async2 = std::async(std::launch::async, publish_data);
-  auto async3 = std::async(std::launch::async, publish_data);
-  auto async4 = std::async(std::launch::async, publish_data);
-  auto async5 = std::async(std::launch::async, publish_data);
-
-  async1.get();
-  async2.get();
-  async3.get();
-  async4.get();
-  async5.get();
-}
-
-TEST_P(StreamLayerClientOnlineTest, ConcurrentPublishDifferentIngestApi) {
-  auto publishData = [&]() {
-    auto client = CreateStreamLayerClient();
-
-    auto response =
-        client
-            ->PublishData(PublishDataRequest().WithData(data_).WithLayerId(
-                GetTestLayer()))
-            .GetFuture()
-            .get();
-
-    ASSERT_NO_FATAL_FAILURE(PublishDataSuccessAssertions(response));
-  };
-
-  auto async1 = std::async(std::launch::async, publishData);
-  auto async2 = std::async(std::launch::async, publishData);
-  auto async3 = std::async(std::launch::async, publishData);
-  auto async4 = std::async(std::launch::async, publishData);
-  auto async5 = std::async(std::launch::async, publishData);
-
-  async1.get();
-  async2.get();
-  async3.get();
-  async4.get();
-  async5.get();
-}
-
-TEST_P(StreamLayerClientOnlineTest, PublishSdii) {
-  auto response = client_
-                      ->PublishSdii(PublishSdiiRequest()
-                                        .WithSdiiMessageList(sdii_data_)
-                                        .WithLayerId(GetTestLayerSdii()))
-                      .GetFuture()
-                      .get();
-
-  ASSERT_NO_FATAL_FAILURE(PublishSdiiSuccessAssertions(response));
-}
-
-TEST_P(StreamLayerClientOnlineTest, PublishSdiiAsync) {
-  std::promise<PublishSdiiResponse> response_promise;
-  bool call_is_async = true;
-  auto cancel_token =
-      client_->PublishSdii(PublishSdiiRequest()
-                               .WithSdiiMessageList(sdii_data_)
-                               .WithLayerId(GetTestLayerSdii()),
-                           [&](const PublishSdiiResponse& response) {
-                             call_is_async = false;
-                             response_promise.set_value(response);
-                           });
-
-  EXPECT_TRUE(call_is_async);
-  auto response_future = response_promise.get_future();
-  auto status = response_future.wait_for(std::chrono::seconds(30));
-  if (status != std::future_status::ready) {
-    cancel_token.cancel();
-  }
-  auto response = response_future.get();
-
-  ASSERT_NO_FATAL_FAILURE(PublishSdiiSuccessAssertions(response));
-}
-
-TEST_P(StreamLayerClientOnlineTest, PublishSdiiCancel) {
-  auto cancel_future =
-      client_->PublishSdii(PublishSdiiRequest()
-                               .WithSdiiMessageList(sdii_data_)
-                               .WithLayerId(GetTestLayerSdii()));
-
-  std::thread([cancel_future]() {
-    std::this_thread::sleep_for(std::chrono::milliseconds(200));
-    cancel_future.GetCancellationToken().cancel();
-  }).detach();
-
-  auto response = cancel_future.GetFuture().get();
-
-  // If the response is successful, do not fail to avoid flakiness. This could
-  // be becauase of a fast network for example.
-  if (response.IsSuccessful()) {
-    SUCCEED();
-    return;
-  }
-
-  ASSERT_NO_FATAL_FAILURE(PublishFailureAssertions(response));
-  // TODO OlpClient/ API level should be updated to fill in the proper APIError
-  // from HttpResponse?
-  //  ASSERT_EQ(olp::client::ErrorCode::Cancelled,
-  //  response.GetError().GetErrorCode());
-}
-
-TEST_P(StreamLayerClientOnlineTest, PublishSdiiCancelLongDelay) {
-  auto cancel_future =
-      client_->PublishSdii(PublishSdiiRequest()
-                               .WithSdiiMessageList(sdii_data_)
-                               .WithLayerId(GetTestLayerSdii()));
-
-  std::thread([cancel_future]() {
-    std::this_thread::sleep_for(std::chrono::milliseconds(1200));
-    cancel_future.GetCancellationToken().cancel();
-  }).detach();
-
-  auto response = cancel_future.GetFuture().get();
-
-  // If the response is successful, do not fail to avoid flakiness. This could
-  // be becauase of a fast network for example.
-  if (response.IsSuccessful()) {
-    SUCCEED();
-    return;
-  }
-
-  ASSERT_NO_FATAL_FAILURE(PublishFailureAssertions(response));
-  // TODO OlpClient/ API level should be updated to fill in the proper APIError
-  // from HttpResponse?
-  //  ASSERT_EQ(olp::client::ErrorCode::Cancelled,
-  //  response.GetError().GetErrorCode());
-}
-
-TEST_P(StreamLayerClientOnlineTest, PublishSDIINonSDIIData) {
-  auto response =
-      client_
-          ->PublishSdii(
-              PublishSdiiRequest().WithSdiiMessageList(data_).WithLayerId(
-                  GetTestLayerSdii()))
-          .GetFuture()
-          .get();
-
-  ASSERT_NO_FATAL_FAILURE(PublishFailureAssertions(response));
-}
-
-TEST_P(StreamLayerClientOnlineTest, PublishSDIIIncorrectLayer) {
-  auto response = client_
-                      ->PublishSdii(PublishSdiiRequest()
-                                        .WithSdiiMessageList(sdii_data_)
-                                        .WithLayerId("BadLayer"))
-                      .GetFuture()
-                      .get();
-
-  ASSERT_NO_FATAL_FAILURE(PublishFailureAssertions(response));
-}
-
-TEST_P(StreamLayerClientOnlineTest, PublishSDIICustomTraceId) {
-  const auto uuid = GenerateRandomUUID();
-
-  auto response = client_
-                      ->PublishSdii(PublishSdiiRequest()
-                                        .WithSdiiMessageList(sdii_data_)
-                                        .WithLayerId(GetTestLayerSdii())
-                                        .WithTraceId(uuid))
-                      .GetFuture()
-                      .get();
-
-  ASSERT_NO_FATAL_FAILURE(PublishSdiiSuccessAssertions(response));
-
-  EXPECT_STREQ(response.GetResult().GetTraceID().GetParentID().c_str(),
-               uuid.c_str());
-}
-
-TEST_P(StreamLayerClientOnlineTest, PublishSDIIBillingTag) {
-  auto response = client_
-                      ->PublishSdii(PublishSdiiRequest()
-                                        .WithSdiiMessageList(sdii_data_)
-                                        .WithLayerId(GetTestLayerSdii())
-                                        .WithBillingTag(kBillingTag))
-                      .GetFuture()
-                      .get();
-
-  ASSERT_NO_FATAL_FAILURE(PublishSdiiSuccessAssertions(response));
-}
-
-#ifdef DATASERVICE_WRITE_HAS_OPENSSL
-TEST_P(StreamLayerClientOnlineTest, SDIIChecksumValid) {
-  const std::string data_string(sdii_data_->begin(), sdii_data_->end());
-  auto checksum = sha256(data_string);
-
-  auto response = client_
-                      ->PublishSdii(PublishSdiiRequest()
-                                        .WithSdiiMessageList(sdii_data_)
-                                        .WithLayerId(GetTestLayerSdii())
-                                        .WithChecksum(checksum))
-                      .GetFuture()
-                      .get();
-
-  ASSERT_NO_FATAL_FAILURE(PublishSdiiSuccessAssertions(response));
-}
-#endif
-
-TEST_P(StreamLayerClientOnlineTest, SDIIChecksumGarbageString) {
-  auto response = client_
-                      ->PublishSdii(PublishSdiiRequest()
-                                        .WithSdiiMessageList(sdii_data_)
-                                        .WithLayerId(GetTestLayerSdii())
-                                        .WithChecksum("GarbageChecksum"))
-                      .GetFuture()
-                      .get();
-
-  ASSERT_NO_FATAL_FAILURE(PublishFailureAssertions(response));
-}
-
-TEST_P(StreamLayerClientOnlineTest, SDIIConcurrentPublishSameIngestApi) {
-  auto publish_data = [&]() {
-    auto response = client_
-                        ->PublishSdii(PublishSdiiRequest()
-                                          .WithSdiiMessageList(sdii_data_)
-                                          .WithLayerId(GetTestLayerSdii()))
-                        .GetFuture()
-                        .get();
-    ASSERT_NO_FATAL_FAILURE(PublishSdiiSuccessAssertions(response));
-  };
-
-  auto async1 = std::async(std::launch::async, publish_data);
-  auto async2 = std::async(std::launch::async, publish_data);
-  auto async3 = std::async(std::launch::async, publish_data);
-  auto async4 = std::async(std::launch::async, publish_data);
-  auto async5 = std::async(std::launch::async, publish_data);
-
-  async1.get();
-  async2.get();
-  async3.get();
-  async4.get();
-  async5.get();
-}
-
 using testing::_;
 
 class StreamLayerClientMockTest : public StreamLayerClientTestBase {
@@ -1214,540 +660,6 @@ TEST_P(StreamLayerClientMockTest, SDIIConcurrentPublishSameIngestApi) {
   async3.get();
   async4.get();
   async5.get();
-}
-
-class StreamLayerClientCacheOnlineTest : public StreamLayerClientOnlineTest {
- protected:
-  static std::shared_ptr<olp::http::Network> s_network;
-
-  static void SetUpTestSuite() {
-    s_network = olp::client::OlpClientSettingsFactory::
-        CreateDefaultNetworkRequestHandler();
-  }
-
-  virtual std::shared_ptr<StreamLayerClient> CreateStreamLayerClient()
-      override {
-    auto network = s_network;
-
-    olp::authentication::Settings authentication_settings;
-    authentication_settings.token_endpoint_url =
-        CustomParameters::getArgument(kEndpoint);
-    authentication_settings.network_request_handler = network;
-
-    olp::authentication::TokenProviderDefault provider(
-        CustomParameters::getArgument(kAppid),
-        CustomParameters::getArgument(kSecret), authentication_settings);
-
-    olp::client::AuthenticationSettings auth_client_settings;
-    auth_client_settings.provider = provider;
-
-    olp::client::OlpClientSettings settings;
-    settings.authentication_settings = auth_client_settings;
-    settings.network_request_handler = network;
-    settings.task_scheduler =
-        olp::client::OlpClientSettingsFactory::CreateDefaultTaskScheduler(1u);
-
-    disk_cache_ = std::make_shared<olp::cache::DefaultCache>();
-    EXPECT_EQ(disk_cache_->Open(),
-              olp::cache::DefaultCache::StorageOpenResult::Success);
-
-    return std::make_shared<StreamLayerClient>(
-        olp::client::HRN{GetTestCatalog()}, settings, disk_cache_,
-        flush_settings_);
-  }
-
-  virtual void TearDown() override {
-    StreamLayerClientTestBase::TearDown();
-    if (disk_cache_) {
-      disk_cache_->Close();
-    }
-  }
-
-  std::shared_ptr<olp::cache::DefaultCache> disk_cache_;
-  FlushSettings flush_settings_;
-};
-
-// Static network instance is necessary as it needs to outlive any created
-// clients. This is a known limitation as triggered send requests capture the
-// network instance inside the callbacks.
-std::shared_ptr<olp::http::Network> StreamLayerClientCacheOnlineTest::s_network;
-
-INSTANTIATE_TEST_SUITE_P(TestCacheOnline, StreamLayerClientCacheOnlineTest,
-                         ::testing::Values(true));
-
-TEST_P(StreamLayerClientCacheOnlineTest, Queue) {
-  auto error = client_->Queue(
-      PublishDataRequest().WithData(data_).WithLayerId(GetTestLayer()));
-
-  ASSERT_FALSE(error) << error.get();
-}
-
-TEST_P(StreamLayerClientCacheOnlineTest, QueueNullData) {
-  auto error = client_->Queue(
-      PublishDataRequest().WithData(nullptr).WithLayerId(GetTestLayer()));
-
-  ASSERT_TRUE(error);
-}
-
-TEST_P(StreamLayerClientCacheOnlineTest, QueueExtraRequestParams) {
-  const auto uuid = GenerateRandomUUID();
-
-  auto error = client_->Queue(PublishDataRequest()
-                                  .WithData(data_)
-                                  .WithLayerId(GetTestLayer())
-                                  .WithTraceId(uuid)
-                                  .WithBillingTag(kBillingTag));
-
-  ASSERT_FALSE(error) << error.get();
-}
-
-#ifdef DATASERVICE_WRITE_HAS_OPENSSL
-TEST_P(StreamLayerClientCacheOnlineTest, QueueWithChecksum) {
-  const std::string data_string(data_->begin(), data_->end());
-  auto checksum = sha256(data_string);
-
-  auto error = client_->Queue(PublishDataRequest()
-                                  .WithData(data_)
-                                  .WithLayerId(GetTestLayer())
-                                  .WithChecksum(checksum));
-
-  ASSERT_FALSE(error) << error.get();
-}
-#endif
-
-TEST_P(StreamLayerClientCacheOnlineTest, FlushDataSingle) {
-  auto error = client_->Queue(
-      PublishDataRequest().WithData(data_).WithLayerId(GetTestLayer()));
-
-  ASSERT_FALSE(error) << error.get();
-
-  auto response = client_->Flush().GetFuture().get();
-
-  ASSERT_FALSE(response.empty());
-  ASSERT_NO_FATAL_FAILURE(PublishDataSuccessAssertions(response[0]));
-}
-
-TEST_P(StreamLayerClientCacheOnlineTest, FlushDataMultiple) {
-  ASSERT_NO_FATAL_FAILURE(QueueMultipleEvents(5));
-
-  auto response = client_->Flush().GetFuture().get();
-
-  ASSERT_EQ(5, response.size());
-  for (auto& single_response : response) {
-    ASSERT_NO_FATAL_FAILURE(PublishDataSuccessAssertions(single_response));
-  }
-}
-
-TEST_P(StreamLayerClientCacheOnlineTest, FlushDataSingleAsync) {
-  auto error = client_->Queue(
-      PublishDataRequest().WithData(data_).WithLayerId(GetTestLayer()));
-
-  ASSERT_FALSE(error) << error.get();
-
-  std::promise<StreamLayerClient::FlushResponse> response_promise;
-  bool call_is_async = true;
-  auto cancel_token =
-      client_->Flush([&](StreamLayerClient::FlushResponse response) {
-        call_is_async = false;
-        response_promise.set_value(response);
-      });
-
-  EXPECT_TRUE(call_is_async);
-  auto response_future = response_promise.get_future();
-  auto status = response_future.wait_for(std::chrono::seconds(30));
-  if (status != std::future_status::ready) {
-    cancel_token.cancel();
-  }
-  auto response = response_future.get();
-
-  ASSERT_FALSE(response.empty());
-  ASSERT_NO_FATAL_FAILURE(PublishDataSuccessAssertions(response[0]));
-}
-
-TEST_P(StreamLayerClientCacheOnlineTest, FlushDataMultipleAsync) {
-  ASSERT_NO_FATAL_FAILURE(QueueMultipleEvents(5));
-
-  std::promise<StreamLayerClient::FlushResponse> response_promise;
-  bool call_is_async = true;
-  auto cancel_token =
-      client_->Flush([&](StreamLayerClient::FlushResponse response) {
-        call_is_async = false;
-        response_promise.set_value(response);
-      });
-
-  EXPECT_TRUE(call_is_async);
-  auto response_future = response_promise.get_future();
-  auto status = response_future.wait_for(std::chrono::seconds(30));
-  if (status != std::future_status::ready) {
-    cancel_token.cancel();
-  }
-  auto response = response_future.get();
-
-  ASSERT_EQ(5, response.size());
-  for (auto& single_response : response) {
-    ASSERT_NO_FATAL_FAILURE(PublishDataSuccessAssertions(single_response));
-  }
-}
-
-TEST_P(StreamLayerClientCacheOnlineTest, FlushDataCancel) {
-  auto error = client_->Queue(
-      PublishDataRequest().WithData(data_).WithLayerId(GetTestLayer()));
-
-  ASSERT_FALSE(error) << error.get();
-
-  auto cancel_future = client_->Flush();
-
-  std::thread([cancel_future]() {
-    std::this_thread::sleep_for(std::chrono::milliseconds(200));
-    cancel_future.GetCancellationToken().cancel();
-  }).detach();
-
-  auto response = cancel_future.GetFuture().get();
-
-  ASSERT_EQ(1, response.size());
-  if (response[0].IsSuccessful()) {
-    SUCCEED();
-    return;
-  }
-
-  ASSERT_NO_FATAL_FAILURE(PublishFailureAssertions(response[0]));
-}
-
-TEST_P(StreamLayerClientCacheOnlineTest, FlushListenerMetrics) {
-  disk_cache_->Close();
-  flush_settings_.auto_flush_num_events = 3;
-  client_ = CreateStreamLayerClient();
-
-  ASSERT_NO_FATAL_FAILURE(QueueMultipleEvents(3));
-
-  auto default_listener = StreamLayerClient::DefaultListener();
-
-  client_->Enable(default_listener);
-
-  for (int i = 0; default_listener->GetNumFlushEvents() < 1; i++) {
-    std::this_thread::sleep_for(std::chrono::milliseconds(50));
-    if (i > 200) {
-      FAIL() << "Timeout waiting for Flush Event Listener Results";
-    }
-  }
-
-  EXPECT_EQ(1, default_listener->GetNumFlushEvents());
-  EXPECT_EQ(1, default_listener->GetNumFlushEventsAttempted());
-  EXPECT_EQ(0, default_listener->GetNumFlushEventsFailed());
-  EXPECT_EQ(3, default_listener->GetNumFlushedRequests());
-  EXPECT_EQ(0, default_listener->GetNumFlushedRequestsFailed());
-}
-
-TEST_P(StreamLayerClientCacheOnlineTest,
-       FlushListenerMetricsSetListenerBeforeQueuing) {
-  disk_cache_->Close();
-  flush_settings_.auto_flush_num_events = 3;
-  client_ = CreateStreamLayerClient();
-
-  auto default_listener = StreamLayerClient::DefaultListener();
-  client_->Enable(default_listener);
-
-  ASSERT_NO_FATAL_FAILURE(QueueMultipleEvents(3));
-
-  for (int i = 0; default_listener->GetNumFlushEvents() < 1; i++) {
-    std::this_thread::sleep_for(std::chrono::milliseconds(50));
-    if (i > 200) {
-      FAIL() << "Timeout waiting for Flush Event Listener Results";
-    }
-  }
-
-  EXPECT_EQ(1, default_listener->GetNumFlushEvents());
-  EXPECT_EQ(1, default_listener->GetNumFlushEventsAttempted());
-  EXPECT_EQ(0, default_listener->GetNumFlushEventsFailed());
-  EXPECT_EQ(3, default_listener->GetNumFlushedRequests());
-  EXPECT_EQ(0, default_listener->GetNumFlushedRequestsFailed());
-}
-
-TEST_P(StreamLayerClientCacheOnlineTest, FlushListenerDisable) {
-  disk_cache_->Close();
-  flush_settings_.auto_flush_num_events = 3;
-  client_ = CreateStreamLayerClient();
-
-  auto default_listener = StreamLayerClient::DefaultListener();
-  client_->Enable(default_listener);
-
-  ASSERT_NO_FATAL_FAILURE(QueueMultipleEvents(3));
-
-  std::this_thread::sleep_for(std::chrono::milliseconds(100));
-
-  auto disable_future = client_->Disable();
-  auto status = disable_future.wait_for(std::chrono::seconds(5));
-  if (status != std::future_status::ready) {
-    FAIL() << "Timeout waiting for auto flushing to be disabled";
-  }
-  disable_future.get();
-
-  EXPECT_EQ(1, default_listener->GetNumFlushEvents());
-  EXPECT_EQ(1, default_listener->GetNumFlushEventsAttempted());
-  EXPECT_EQ(1, default_listener->GetNumFlushEventsFailed());
-}
-
-TEST_P(StreamLayerClientCacheOnlineTest,
-       FlushListenerMetricsMultipleFlushEventsInSeries) {
-  disk_cache_->Close();
-  flush_settings_.auto_flush_num_events = 2;
-  client_ = CreateStreamLayerClient();
-
-  auto default_listener = StreamLayerClient::DefaultListener();
-  client_->Enable(default_listener);
-
-  ASSERT_NO_FATAL_FAILURE(QueueMultipleEvents(2));
-
-  for (int i = 0, j = 1;; i++) {
-    std::this_thread::sleep_for(std::chrono::milliseconds(50));
-    if (default_listener->GetNumFlushEvents() == j) {
-      if (j == 3) {
-        break;
-      }
-      ASSERT_NO_FATAL_FAILURE(QueueMultipleEvents(2));
-      j++;
-    }
-    if (i > 400) {
-      FAIL() << "Timeout waiting for Flush Event Listener Results";
-    }
-  }
-
-  EXPECT_EQ(3, default_listener->GetNumFlushEvents());
-  EXPECT_EQ(3, default_listener->GetNumFlushEventsAttempted());
-  EXPECT_EQ(0, default_listener->GetNumFlushEventsFailed());
-  EXPECT_EQ(6, default_listener->GetNumFlushedRequests());
-  EXPECT_EQ(0, default_listener->GetNumFlushedRequestsFailed());
-}
-
-TEST_P(StreamLayerClientCacheOnlineTest,
-       FlushListenerMetricsMultipleFlushEventsInParallel) {
-  disk_cache_->Close();
-  flush_settings_.auto_flush_num_events = 2;
-  flush_settings_.events_per_single_flush =
-      flush_settings_.auto_flush_num_events;
-  client_ = CreateStreamLayerClient();
-
-  auto default_listener = StreamLayerClient::DefaultListener();
-  client_->Enable(default_listener);
-
-  ASSERT_NO_FATAL_FAILURE(QueueMultipleEvents(6));
-
-  for (int i = 0; default_listener->GetNumFlushedRequests() < 6; i++) {
-    std::this_thread::sleep_for(std::chrono::milliseconds(50));
-    if (i > 500) {
-      FAIL() << "Timeout waiting for Flush Event Listener Results";
-    }
-  }
-
-  EXPECT_LE(3, default_listener->GetNumFlushEvents());
-  EXPECT_LE(3, default_listener->GetNumFlushEventsAttempted());
-  // TODO: Investigate why there are more triggeres in auto flushing then
-  // requests. Seems AutoFlushController is trying to flush to often.
-  // EXPECT_EQ(0, default_listener->GetNumFlushEventsFailed());
-  EXPECT_EQ(6, default_listener->GetNumFlushedRequests());
-  EXPECT_EQ(0, default_listener->GetNumFlushedRequestsFailed());
-}
-
-TEST_P(StreamLayerClientCacheOnlineTest,
-       FlushListenerMetricsMultipleFlushEventsInParallelStaggeredQueue) {
-  disk_cache_->Close();
-  flush_settings_.auto_flush_num_events = 2;
-  flush_settings_.events_per_single_flush =
-      flush_settings_.auto_flush_num_events;
-  client_ = CreateStreamLayerClient();
-
-  auto default_listener = StreamLayerClient::DefaultListener();
-  client_->Enable(default_listener);
-
-  ASSERT_NO_FATAL_FAILURE(QueueMultipleEvents(4));
-  std::this_thread::sleep_for(std::chrono::milliseconds(50));
-  ASSERT_NO_FATAL_FAILURE(QueueMultipleEvents(2));
-  std::this_thread::sleep_for(std::chrono::milliseconds(100));
-  ASSERT_NO_FATAL_FAILURE(QueueMultipleEvents(4));
-
-  for (int i = 0; default_listener->GetNumFlushedRequests() < 10; i++) {
-    std::this_thread::sleep_for(std::chrono::milliseconds(50));
-    if (i > 600) {
-      FAIL() << "Timeout waiting for Flush Event Listener Results";
-    }
-  }
-
-  EXPECT_LE(3, default_listener->GetNumFlushEvents());
-  EXPECT_LE(3, default_listener->GetNumFlushEventsAttempted());
-  // TODO: Investigate why there are more triggeres in auto flushing then
-  // requests. Seems AutoFlushController is trying to flush to often.
-  // EXPECT_EQ(0, default_listener->GetNumFlushEventsFailed());
-  EXPECT_EQ(10, default_listener->GetNumFlushedRequests());
-  EXPECT_EQ(0, default_listener->GetNumFlushedRequestsFailed());
-}
-
-TEST_P(StreamLayerClientCacheOnlineTest, FlushListenerNotifications) {
-  disk_cache_->Close();
-  flush_settings_.auto_flush_num_events = 3;
-  client_ = CreateStreamLayerClient();
-
-  ASSERT_NO_FATAL_FAILURE(QueueMultipleEvents(3));
-
-  class NotificationListener : public DefaultFlushEventListener<
-                                   const StreamLayerClient::FlushResponse&> {
-   public:
-    void NotifyFlushEventStarted() override { events_started_++; }
-
-    void NotifyFlushEventResults(
-        const StreamLayerClient::FlushResponse& results) override {
-      std::lock_guard<std::mutex> lock(results_mutex_);
-      results_ = std::move(results);
-    }
-
-    const StreamLayerClient::FlushResponse& GetResults() {
-      std::lock_guard<std::mutex> lock(results_mutex_);
-      return results_;
-    }
-
-    int events_started_ = 0;
-
-   private:
-    StreamLayerClient::FlushResponse results_;
-    std::mutex results_mutex_;
-  };
-
-  auto notification_listener = std::make_shared<NotificationListener>();
-  client_->Enable(notification_listener);
-
-  for (int i = 0; notification_listener->GetResults().size() < 3; i++) {
-    std::this_thread::sleep_for(std::chrono::milliseconds(50));
-    if (i > 200) {
-      FAIL() << "Timeout waiting for Flush Event Listener Results";
-    }
-  }
-
-  EXPECT_EQ(1, notification_listener->events_started_);
-  for (auto result : notification_listener->GetResults()) {
-    ASSERT_NO_FATAL_FAILURE(PublishDataSuccessAssertions(result));
-  }
-}
-
-TEST_P(StreamLayerClientCacheOnlineTest, FlushSettingsTimeSinceOldRequest) {
-  disk_cache_->Close();
-  flush_settings_.auto_flush_old_events_force_flush_interval = 10;
-  client_ = CreateStreamLayerClient();
-
-  ASSERT_NO_FATAL_FAILURE(QueueMultipleEvents(2));
-
-  auto default_listener = StreamLayerClient::DefaultListener();
-  client_->Enable(default_listener);
-
-  for (int i = 0; default_listener->GetNumFlushEvents() < 1; i++) {
-    std::this_thread::sleep_for(std::chrono::milliseconds(50));
-    if (i > 400) {
-      FAIL() << "Timeout waiting for Flush Event Listener Results";
-    }
-  }
-
-  EXPECT_EQ(1, default_listener->GetNumFlushEvents());
-  EXPECT_EQ(1, default_listener->GetNumFlushEventsAttempted());
-  EXPECT_EQ(0, default_listener->GetNumFlushEventsFailed());
-  EXPECT_EQ(2, default_listener->GetNumFlushedRequests());
-  EXPECT_EQ(0, default_listener->GetNumFlushedRequestsFailed());
-}
-
-TEST_P(StreamLayerClientCacheOnlineTest,
-       FlushSettingsTimeSinceOldRequestQueueAfterEnable) {
-  disk_cache_->Close();
-  flush_settings_.auto_flush_old_events_force_flush_interval = 10;
-  client_ = CreateStreamLayerClient();
-
-  auto default_listener = StreamLayerClient::DefaultListener();
-  client_->Enable(default_listener);
-
-  ASSERT_NO_FATAL_FAILURE(QueueMultipleEvents(2));
-
-  for (int i = 0; default_listener->GetNumFlushEvents() < 1; i++) {
-    std::this_thread::sleep_for(std::chrono::milliseconds(50));
-    if (i > 400) {
-      FAIL() << "Timeout waiting for Flush Event Listener Results";
-    }
-  }
-
-  EXPECT_EQ(1, default_listener->GetNumFlushEvents());
-  EXPECT_EQ(1, default_listener->GetNumFlushEventsAttempted());
-  EXPECT_EQ(0, default_listener->GetNumFlushEventsFailed());
-  EXPECT_EQ(2, default_listener->GetNumFlushedRequests());
-  EXPECT_EQ(0, default_listener->GetNumFlushedRequestsFailed());
-}
-
-TEST_P(StreamLayerClientCacheOnlineTest,
-       FlushSettingsTimeSinceOldRequestDisable) {
-  disk_cache_->Close();
-  flush_settings_.auto_flush_old_events_force_flush_interval = 2;
-  client_ = CreateStreamLayerClient();
-
-  ASSERT_NO_FATAL_FAILURE(QueueMultipleEvents(2));
-
-  auto default_listener = StreamLayerClient::DefaultListener();
-  client_->Enable(default_listener);
-
-  std::this_thread::sleep_for(std::chrono::milliseconds(2100));
-
-  auto disable_future = client_->Disable();
-  auto status = disable_future.wait_for(std::chrono::seconds(5));
-  if (status != std::future_status::ready) {
-    FAIL() << "Timeout waiting for auto flushing to be disabled";
-  }
-  disable_future.get();
-
-  EXPECT_EQ(1, default_listener->GetNumFlushEvents());
-  EXPECT_EQ(1, default_listener->GetNumFlushEventsAttempted());
-  EXPECT_EQ(1, default_listener->GetNumFlushEventsFailed());
-}
-
-TEST_P(StreamLayerClientCacheOnlineTest, FlushSettingsAutoFlushInterval) {
-  disk_cache_->Close();
-  flush_settings_.auto_flush_interval = 10;
-  client_ = CreateStreamLayerClient();
-
-  ASSERT_NO_FATAL_FAILURE(QueueMultipleEvents(2));
-
-  auto default_listener = StreamLayerClient::DefaultListener();
-  client_->Enable(default_listener);
-
-  for (int i = 0; default_listener->GetNumFlushEvents() < 1; i++) {
-    std::this_thread::sleep_for(std::chrono::milliseconds(50));
-    if (i > 400) {
-      FAIL() << "Timeout waiting for Flush Event Listener Results";
-    }
-  }
-
-  EXPECT_EQ(1, default_listener->GetNumFlushEvents());
-  EXPECT_EQ(1, default_listener->GetNumFlushEventsAttempted());
-  EXPECT_EQ(0, default_listener->GetNumFlushEventsFailed());
-  EXPECT_EQ(2, default_listener->GetNumFlushedRequests());
-  EXPECT_EQ(0, default_listener->GetNumFlushedRequestsFailed());
-}
-
-TEST_P(StreamLayerClientCacheOnlineTest,
-       FlushSettingsAutoFlushIntervalDisable) {
-  disk_cache_->Close();
-  flush_settings_.auto_flush_interval = 2;
-  client_ = CreateStreamLayerClient();
-
-  ASSERT_NO_FATAL_FAILURE(QueueMultipleEvents(2));
-
-  auto default_listener = StreamLayerClient::DefaultListener();
-  client_->Enable(default_listener);
-
-  std::this_thread::sleep_for(std::chrono::milliseconds(2100));
-
-  auto disable_future = client_->Disable();
-  auto status = disable_future.wait_for(std::chrono::seconds(5));
-  if (status != std::future_status::ready) {
-    FAIL() << "Timeout waiting for auto flushing to be disabled";
-  }
-  disable_future.get();
-
-  EXPECT_EQ(1, default_listener->GetNumFlushEvents());
-  EXPECT_EQ(1, default_listener->GetNumFlushEventsAttempted());
-  EXPECT_EQ(1, default_listener->GetNumFlushEventsFailed());
 }
 
 class StreamLayerClientCacheMockTest : public StreamLayerClientMockTest {
